requirements.txt
gspread>=6.2.1
google-auth>=2.30.0
